+++ server/statistics/store_collection_test.go
